+++ src/views/OnboardingView/CompleteProfile/index.vue
@@ -278,7 +278,7 @@ import validator from 'validator'
 import UserService from 'src/services/UserService'
 import OnboardingService from 'src/services/OnboardingService'
 
-import AutocompleteInput from '../molecules/AutocompleteInput'
+import AutocompleteInput from './AutocompleteInput'
 
 export default {
   components: {
